interrface with network card AMD 79c973 initialize sucessfully

+++ include/memorymanagement.h
@@ -28,12 +28,12 @@ namespace myos
     };
 }
 
-void* operator new(size_t size);
-void* operator new[](size_t size);
+void* operator new(unsigned size);
+void* operator new[](unsigned size);
 
 //placement new
-void* operator new(size_t size, void* ptr);
-void* operator new[](size_t size, void* ptr);
+void* operator new(unsigned size, void* ptr);
+void* operator new[](unsigned size, void* ptr);
 
 void operator delete(void* ptr);
 void operator delete[](void* ptr);
